requirements.txt
dj-database-url==0.5.0
Django==3.0.4
django-nose==1.4.6
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-nested-routers==0.91
psycopg2-binary==2.8.4
pyyaml==5.3.1
requests==2.23.0
uritemplate==3.0.1
uuid==1.30.0
whitenoise==5.0.1